Setup support for serverless
