--- .drone.yml
@@ -1,20 +1,0 @@
-kind: pipeline
-name: default
-
-steps:
-  - name: deploy
-    image: appleboy/drone-ssh
-    pull: true
-    secrets: [ssh_key]
-    when:
-      event: push
-      branch: master
-    settings:
-      host: nansen.online.ntnu.no
-      port: 22
-      username: root
-      key:
-        from_secret: ssh_key
-      command_timeout: 600s
-      script:
-        - /srv/www/smashrank/deploy.sh
